Set up the locationConfig, localStorage, api and style of app, and search area component
Previously, the search area and gallery component have been
created.
This commit ajusts the style of the app and search component.
This commit also adds the api file contains the asynchronous
call for fetching the pictures.
This commit also sets up the location config file for setting
the default options for selection.
This commit adds the storageUtils file to store the data in
the local storage.