Set up the redux
We need to use redux to manage the reusable and global
state for sharing behaviour between different components.
This commit adds the action-types for setting the constants
for actions.
This commit adds the actions to define different actions by
action creators, like setting the latitude, longtitude, get
photo data by asynchronous operation and so on.
This commit adds the reducer to handle the actions by previous
state and actions.
This commit also set up the store which holds the state tree
of the application.